requirements.txt
langchain-openai
langchain
python-dotenv
langchain_community
langchain_core
streamlit
ipykernel
pypdf
bs4
arxiv
pymupdf
wikipedia
langchain-text-splitters
chromadb
sentence_transformers
langchain_huggingface
faiss-cpu
langchain_chroma
duckdb
openai
langchain-groq
langchain_ollama
fastapi
uvicorn
python-multipart
db-sqlite3
langserve
sse_starlette
ace-tools
mlflow